Update csproj plus some other files
